requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Authlib==1.3.0
python-dotenv==1.0.0
requests==2.31.0
cloudinary==1.36.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
